+++ core/src/test/java/com/cobber/fta/DetermineDateTimeFormatTests.java
@@ -1467,7 +1467,7 @@ public class DetermineDateTimeFormatTests {
 	public void bogusInput() {
 		final String pipedInput = "12:45:64.|21/12/99:|21/12/99:|18:46:|4:38  39|3124/08/|890/65 1/|7/87/33| 89:50|18:52 56:|18/94/06|0463 5 71|50 9:22|" +
 				"95/06/88|0-27-98|08/56 22/|31-0-99|0/7:6/11 //61|8:73/4/13 15|14/23/3367| 00/21/79|22-23-00|0/20/2361|0/2/52 9:50 4 |" +
-				"2015-8-17T|4/01/41 3:43 T450|37/8/005 5:05|0/6/95|0000 7 1|2000-12-12T12:45-72|2000-12-12T12:45-112|" +
+				"4/01/41 3:43 T450|37/8/005 5:05|0/6/95|2000-12-12T12:45-72|2000-12-12T12:45-112|" +
 				"12:45:64.|84:12:45.5712:45| 12:45:63.3 |";
 		final String[] inputs = pipedInput.split("\\|");
 
@@ -1475,6 +1475,8 @@ public class DetermineDateTimeFormatTests {
 			final DateTimeParser det = new DateTimeParser();
 			det.train(testCase);
 			final DateTimeParserResult result = det.getResult();
+			if (result != null)
+				System.err.println(result.getFormatString());
 			assertNull(result, testCase);
 		}
 	}
@@ -1955,6 +1957,62 @@ public class DetermineDateTimeFormatTests {
 		assertFalse(result.isValid8("118:03:59"));
 		assertFalse(result.isValid8("118:3:59"));
 		assertFalse(result.isValid8("118:333:59"));
+	}
+
+	@Test(groups = { TestGroups.ALL, TestGroups.DATETIME })
+	public void intuitMMMHHmmssa() {
+		final DateTimeParser dtp = new DateTimeParser(DateResolutionMode.None, Locale.getDefault());
+
+		assertEquals(dtp.determineFormatString("May 8, 2009 5:57:51 PM"), "MMM d, yyyy h:mm:ss a");
+	}
+
+	@Test(groups = { TestGroups.ALL, TestGroups.DATETIME })
+	public void intuityyyyMMddHmm() {
+		final DateTimeParser dtp = new DateTimeParser(DateResolutionMode.None, Locale.getDefault());
+
+		assertEquals(dtp.determineFormatString("2014/4/8 22:05"), "yyyy/M/d HH:mm");
+	}
+
+	@Test(groups = { TestGroups.ALL, TestGroups.DATETIME })
+	public void embeddedQuote() {
+		final DateTimeParser dtp = new DateTimeParser(DateResolutionMode.None, Locale.getDefault());
+
+		assertEquals(dtp.determineFormatString("oct 7, '70"), "MMM d, ''yy");
+	}
+
+	@Test(groups = { TestGroups.ALL, TestGroups.DATETIME })
+	public void intuitMMMdyy() {
+		final DateTimeParser dtp = new DateTimeParser(DateResolutionMode.None, Locale.getDefault());
+
+		assertEquals(dtp.determineFormatString("oct. 7, 70"), "MMM. d, yy");
+	}
+
+	@Test(groups = { TestGroups.ALL, TestGroups.DATETIME })
+	public void intuityyyyMMdHHmmss() {
+		final DateTimeParser dtp = new DateTimeParser(DateResolutionMode.None, Locale.getDefault());
+
+		assertEquals(dtp.determineFormatString("2014/04/2 03:00:51"), "yyyy/MM/d HH:mm:ss");
+	}
+
+	@Test(groups = { TestGroups.ALL, TestGroups.DATETIME })
+	public void intuityyyyMdHHmm() {
+		final DateTimeParser dtp = new DateTimeParser(DateResolutionMode.None, Locale.getDefault());
+
+		assertEquals(dtp.determineFormatString("2014:4:8 22:05"), "yyyy:M:d HH:mm");
+	}
+
+	@Test(groups = { TestGroups.ALL, TestGroups.DATETIME })
+	public void intuityyyyMMdHHmmssColons() {
+		final DateTimeParser dtp = new DateTimeParser(DateResolutionMode.None, Locale.getDefault());
+
+		assertEquals(dtp.determineFormatString("2014:04:2 03:00:51"), "yyyy:MM:d HH:mm:ss");
+	}
+
+	@Test(groups = { TestGroups.ALL, TestGroups.DATETIME })
+	public void intuityyyyMddHHmmssColons() {
+		final DateTimeParser dtp = new DateTimeParser(DateResolutionMode.None, Locale.getDefault());
+
+		assertEquals(dtp.determineFormatString("2014:4:02 03:00:51"), "yyyy:M:dd HH:mm:ss");
 	}
 
 	@Test(groups = { TestGroups.ALL, TestGroups.DATETIME })